Fix/ ArenaResultResponse: 순위, 참여한 사람 변수형 long으로 변경

+++ src/main/java/com/example/KeyboardArenaProject/dto/arena/ArenaResultResponse.java
@@ -14,11 +14,11 @@ import java.time.LocalTime;
 public class ArenaResultResponse {
     String title;
     LocalTime time;
-    Integer participants;
-    Integer nthPlace;
+    Long participants;
+    Long nthPlace;
     Double percentage;
 
-    @Builder ArenaResultResponse(String title, int participants, LocalTime time, int nthPlace) {
+    @Builder ArenaResultResponse(String title, Long participants, LocalTime time, Long nthPlace) {
         this.title = title;
         this.time = time;
         this.participants = participants;
